Move static stitch code to helper class
