+++ src/Features/LanguageServer/ProtocolUnitTests/SemanticTokens/SemanticTokensRangesTests.cs
@@ -4,11 +4,9 @@
 
 #nullable enable
 
-using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Classification;
-using Microsoft.CodeAnalysis.LanguageServer.ExternalAccess.Razor;
 using Microsoft.CodeAnalysis.LanguageServer.Handler.SemanticTokens;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 using Roslyn.Test.Utilities;
@@ -69,61 +67,6 @@ static class C { }
 
             await VerifyBasicInvariantsAndNoMultiLineTokens(testLspServer, results.Data).ConfigureAwait(false);
             AssertEx.Equal(ConvertToReadableFormat(testLspServer.ClientCapabilities, expectedResults.Data), ConvertToReadableFormat(testLspServer.ClientCapabilities, results.Data));
-        }
-
-        [Fact]
-        public void StitchSemanticTokenResponsesTogether_OnEmptyInput_ReturnsEmptyResponseData()
-        {
-            // Arrange
-            var responseData = Array.Empty<int[]>();
-
-            // Act
-            var result = SemanticTokensRangesHandler.StitchSemanticTokenResponsesTogether(responseData);
-
-            // Assert
-            Assert.NotNull(result);
-            Assert.Empty(result);
-        }
-
-        [Fact]
-        public void StitchSemanticTokenResponsesTogether_ReturnsCombinedResponseData()
-        {
-            // Arrange
-            var responseData = new int[][] {
-             new int[] { 0, 0, 0, 0, 0,
-                         1, 0, 0, 0, 0,
-                         1, 0, 0, 0, 0,
-                         0, 5, 0, 0, 0,
-                         0, 3, 0, 0, 0,
-                         2, 2, 0, 0, 0,
-                         0, 3, 0, 0, 0 },
-             new int[] { 10, 0, 0, 0, 0,
-                         1, 0, 0, 0, 0,
-                         1, 0, 0, 0, 0,
-                         0, 5, 0, 0, 0,
-                         0, 3, 0, 0, 0,
-                         2, 2, 0, 0, 0,
-                         0, 3, 0, 0, 0 },
-             new int[] { 14, 7, 0, 0, 0,
-                         1, 0, 0, 0, 0,
-                         1, 0, 0, 0, 0,
-                         0, 5, 0, 0, 0,
-                         0, 3, 0, 0, 0,
-                         2, 2, 0, 0, 0,
-                         0, 3, 0, 0, 0 },
-         };
-
-            var expectedResponseData = new int[] {
-            0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 5, 0, 0, 0, 0, 3, 0, 0, 0, 2, 2, 0, 0, 0, 0, 3, 0, 0, 0,
-            6, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 5, 0, 0, 0, 0, 3, 0, 0, 0, 2, 2, 0, 0, 0, 0, 3, 0, 0, 0,
-            0, 2, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 5, 0, 0, 0, 0, 3, 0, 0, 0, 2, 2, 0, 0, 0, 0, 3, 0, 0, 0
-        };
-
-            // Act
-            var result = SemanticTokensRangesHandler.StitchSemanticTokenResponsesTogether(responseData);
-
-            // Assert
-            Assert.Equal(expectedResponseData, result);
         }
     }
 }
